View troubleshooting joe

+++ src/main/java/com/codeup/phaserun/repositories/RaceRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 @Repository
 public interface RaceRepository extends JpaRepository<Race, Long> {
-    Race findById(int id);
+    Race findById(long id);
     boolean existsByRaceId(String raceId);
     Race findByRaceId(String raceId);
 }
